WebIDL/tests/submissions/heycam/ArrayClass-001.html

<!DOCTYPE html>
<title>[ArrayClass]</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.5.3 Interface prototype object
//
// [ArrayClass] on an interface means that its [[Prototype]] is Array.prototype.

// feature: NodeList (interface) (with [ArrayClass])

test(function() {
  assert_equals(NodeList.prototype, Array.prototype, "NodeList.prototype == Array.prototype");
}, "NodeList being [ArrayClass] implies NodeList.prototype == Array.prototype");

test(function() {
  var a = document.childNodes.slice();
  assert_equals(Object.getPrototypeOf(a), Array.prototype, "nodeList.slice() returns an Array");
  assert_array_equals(document.childNodes, a);
}, "nodeList.slice() works");
</script>
